feat: performance and stability updates

+++ src/server/api/routers/items.ts
@@ -2,7 +2,6 @@ import { z } from "zod";
 import { createTRPCRouter, protectedProcedure } from "../trpc";
 import { TRPCError } from "@trpc/server";
 import { PERMS } from "../../../_constants/perms"
-import { $Enums } from "@prisma/client";
 import { isAuthorised } from "~/utils";
 import { createId } from '@paralleldrive/cuid2';
 
@@ -28,25 +27,7 @@ const addItemSchema = z.object({
     volume: z.number().optional(),
 });
 
-const itemSellSchema = z.object({
-    storageId: nonEmptyString,
-    customerId: nonEmptyString,
-    discount: z.number(),
-    totalPayAmount: z.number(),
-    priceToPay: z.number(),
-    selectedPriceType: z.nativeEnum($Enums.PriceType).optional(),
-    exchangeRate: z.number().optional(),
-    transferToDealer: z.boolean(),
-    paidAmount: z.number(),
-    saleCancel: z.boolean(),
-    items: z.array(z.object({
-        itemId: z.string(),
-        price: z.number().optional(),
-        barcode: nonEmptyString,
-        totalAdded: z.number(),
-        serialNumbers: z.string().array()
-    }))
-})
+
 
 export const itemsRouter = createTRPCRouter({
     getItems: protectedProcedure
@@ -300,23 +281,31 @@ export const itemsRouter = createTRPCRouter({
                 });
             }
 
-            const { data: dealerStorages, error: dealerStoragesError } = await ctx.supabase.from("Storage").select("id,name").eq("orgId", item.orgId ?? "")
+            const { data: dealerStorages, error: dealerStoragesError } = await ctx.supabase.from("Storage").select("id,name,stocks:ItemStock(stock)").eq("orgId", item.orgId ?? "")
             if (dealerStoragesError) {
                 throw new TRPCError({
                     code: "BAD_REQUEST",
                     message: "Failed to get dealer storage",
                 });
             }
-            return dealerStorages
+            const storages = dealerStorages.map(({ stocks, ...rest }) => {
+                const totalStock = stocks.reduce((sum, stock) => sum + stock.stock, 0);
+                return { ...rest, totalStock };
+            });
+            return storages
         }
-        const { data: orgStorages, error: orgStoragesError } = await ctx.supabase.from("Storage").select("id,name").eq("orgId", ctx.session.orgId)
+        const { data: orgStorages, error: orgStoragesError } = await ctx.supabase.from("Storage").select("id,name,stocks:ItemStock(stock)").eq("orgId", ctx.session.orgId)
         if (orgStoragesError) {
             throw new TRPCError({
                 code: "BAD_REQUEST",
                 message: "Failed to get org storage",
             });
         }
-        return orgStorages
+        const storages = orgStorages.map(({ stocks, ...rest }) => {
+            const totalStock = stocks.reduce((sum, stock) => sum + stock.stock, 0);
+            return { ...rest, totalStock };
+        });
+        return storages
 
     }),
 
@@ -354,7 +343,7 @@ export const itemsRouter = createTRPCRouter({
         .mutation(async ({ ctx, input }) => {
             const userPerms = ctx.session.permissions
 
-            if (ctx.session.orgId) {
+            if (!ctx.session.orgId) {
                 throw new TRPCError({
                     code: "UNAUTHORIZED",
                     message: "You don't have permission to do this!",
@@ -364,6 +353,19 @@ export const itemsRouter = createTRPCRouter({
                 throw new TRPCError({
                     code: "UNAUTHORIZED",
                     message: "You don't have permission to do this!",
+                });
+            }
+            const { data: storage, error: getStorageError } = await ctx.supabase.from("Storage").select("*,ItemStock(Item(*))").eq("id", input).single()
+            if (getStorageError) {
+                throw new TRPCError({
+                    code: "UNAUTHORIZED",
+                    message: "You don't have permission to do this!",
+                });
+            }
+            if (storage.ItemStock.length > 0) {
+                throw new TRPCError({
+                    code: "BAD_REQUEST",
+                    message: "There is already an item in this storage!",
                 });
             }
             const { error } = await ctx.supabase.from("Storage").delete().eq("id", input)
@@ -405,8 +407,8 @@ export const itemsRouter = createTRPCRouter({
                     isServiceItem: input.isServiceItem,
                     itemColorId: input.itemColorId,
                     itemSizeId: input.itemSizeId,
-                    netWeight: input.netWeight,
-                    volume: input.volume,
+                    netWeight: input.netWeight ?? 0,
+                    volume: input.volume ?? 0,
                     itemCategoryId: input.itemCategoryId,
                     updateDate: new Date().toUTCString(),
                     createDate: new Date().toUTCString()
@@ -921,7 +923,6 @@ export const itemsRouter = createTRPCRouter({
             });
         }
         if (input.isMaster && masterBarcode && barcode.id !== masterBarcode.id) {
-            console.log("amk hadi ya");
 
             const { error } = await ctx.supabase
                 .from('itemBarcode')
@@ -1000,242 +1001,8 @@ export const itemsRouter = createTRPCRouter({
 
         return "success"
     }),
-    itemAccept: protectedProcedure
-        .input(
-            z.object({
-                storageId: nonEmptyString,
-                fromCustomerId: nonEmptyString,
-                items: z.object({
-                    itemId: nonEmptyString,
-                    barcode: nonEmptyString,
-                    quantity: z.number().min(1),
-                }).array().min(1)
-            }))
-        .mutation(async ({ ctx, input }) => {
-            if (!input) {
-                new TRPCError({
-                    code: "BAD_REQUEST",
-                    message: "Invalid Payload!"
-                })
-            }
-            const userPerms = ctx.session.permissions
 
-            if (!userPerms.includes(PERMS.item_accept)) {
-                throw new TRPCError({
-                    code: "UNAUTHORIZED",
-                    message: "You don't have permission to do this!",
-                });
-            }
-            const { data: storage, error: storageError } = await ctx.supabase.from("Storage").select("*,ItemStock(Item(*))").eq("id", input.storageId).single()
-            if (storageError) {
-                throw new TRPCError({
-                    code: "UNAUTHORIZED",
-                    message: "You don't have permission to do this!",
-                });
-            }
-            if (!storage) {
-                throw new TRPCError({
-                    code: "UNAUTHORIZED",
-                    message: "You don't have permission to do this!",
-                });
-            }
-            input.items.map(async (i) => {
-                const { data: existingItemStock, error: itemStockError } = await ctx.supabase.from("ItemStock").select("*,Item(*)").eq("itemId", i.itemId).eq("storageId", input.storageId).single()
-                if (itemStockError) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Depo bulunamadı!",
-                    });
-                }
-                const { data: barcodeDetails, error: barcodeError } = await ctx.supabase.from("itemBarcode").select("*,Item(*)").eq("barcode", i.barcode).eq("itemId", i.itemId).single()
-                if (barcodeError) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Barkod Bulunamadı!",
-                    });
-                }
-                if (existingItemStock) {
-                    const { error } = await ctx.supabase.from("ItemStock").update({ stock: (i.quantity * barcodeDetails.quantity) + existingItemStock.stock }).eq("id", existingItemStock.id)
-                    if (error) {
-                        throw new TRPCError({
-                            code: "BAD_REQUEST",
-                            message: error.message,
-                        });
-                    }
-                } else {
-                    const { error } = await ctx.supabase.from("ItemStock").insert({ id: createId(), itemId: i.itemId, stock: i.quantity * barcodeDetails.quantity, storageId: input.storageId })
-                    if (error) {
-                        throw new TRPCError({
-                            code: "BAD_REQUEST",
-                            message: error.message,
-                        });
-                    }
-                }
-                const { data: history, error } = await ctx.supabase.from("ItemAcceptHistory")
-                    .insert({
-                        id: createId(),
-                        customerId: input.fromCustomerId,
-                        storageId: input.storageId,
-                        name: ctx.session.email ?? "Bilinmeyen Kullanıcı",
-                        orgId: ctx.session.orgId,
-                        createDate: new Date().toUTCString()
-                    }).select().single()
-                if (error) {
-                    throw new TRPCError({
-                        code: "INTERNAL_SERVER_ERROR",
-                        message: "Internal Server Error",
-                    });
-                }
-                await ctx.supabase.from("ItemAcceptDetail")
-                    .insert({
-                        id: createId(),
-                        itemId: i.itemId,
-                        itemAcceptHistoryId: history.id,
-                        itemBarcodeId: barcodeDetails.id,
-                        quantity: i.quantity,
-                    })
-            })
-            return "success"
-        }),
-    itemSell: protectedProcedure.input(itemSellSchema).mutation(async ({ input, ctx }) => {
-        const userPerms = ctx.session.permissions
 
-        if (!userPerms.includes(PERMS.item_sell)) {
-
-        }
-        const { data: customer, error } = await ctx.supabase.from("Customer").select("*").eq("id", input.customerId).single()
-        if (error) {
-            throw new TRPCError({
-                code: "BAD_REQUEST",
-                message: "Failed to get customer",
-            });
-        }
-        const { data: storage, error: storageError } = await ctx.supabase.from("Storage").select("*").eq("id", input.storageId).single()
-        if (storageError) {
-            throw new TRPCError({
-                code: "BAD_REQUEST",
-                message: "Failed to get storage",
-            });
-        }
-
-        const priceEnum = z.nativeEnum($Enums.PriceType);
-        const priceType = priceEnum.parse(input.selectedPriceType);
-
-        if (!input.saleCancel) {
-            input.items.map(async (i) => {
-                const itemStock = await ctx.supabase.from("ItemStock").select("*").eq("itemId", i.itemId).eq("storageId", storage.id).single()
-                if (itemStock.error) {
-                    throw new TRPCError({
-                        code: "UNAUTHORIZED",
-                        message: "You don't have permission to do this!",
-                    });
-                }
-                console.log(itemStock.data);
-
-                const remainingStock = itemStock.data.stock - i.totalAdded
-                if (remainingStock < 0) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Deponuzdaki Ürünler Yeterli Değil",
-                    });
-                }
-                const { error } = await ctx.supabase.from("ItemStock").update({ stock: remainingStock }).eq("id", itemStock.data?.id)
-                console.log(error);
-
-                if (error) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Failed to update stock " + error.message,
-                    });
-                }
-            })
-        }
-        const member = await ctx.supabase.from("Member").select("*").eq("userEmail", ctx.session.email ?? "").maybeSingle()
-        const transaction = await ctx.supabase.from("Transaction").insert({
-            id: createId(),
-            updatedAt: new Date().toUTCString(),
-            memberId: member.data?.id,
-            orgId: ctx.session.orgId,
-            transferredDealerId: input.transferToDealer ? customer.connectedDealerId : null,
-            customerId: customer.id,
-            discount: String(input.discount),
-            exchangeRate: String(input.exchangeRate),
-            priceType: priceType,
-            storageId: input.storageId,
-            totalAmount: String(input.totalPayAmount),
-            transactionType: input.saleCancel ? "Cancel" : "Sale",
-            payAmount: String(input.paidAmount),
-        }).select().single()
-
-        if (transaction.error) {
-            throw new TRPCError({
-                code: "BAD_REQUEST",
-                message: "Failed to create transaction",
-            })
-        }
-        const createdTransactionItemDetails = await Promise.all(input.items.map(async (i) => {
-            const item = await ctx.supabase.from("Item").select("*").eq("id", i.itemId).single()
-            if (item.error) {
-                throw new TRPCError({
-                    code: "BAD_REQUEST",
-                    message: "Failed to get item",
-                })
-            }
-            if (item.error) {
-                throw new TRPCError({
-                    code: "BAD_REQUEST",
-                    message: "Failed to get item",
-                })
-            }
-            const customerPriceType = customer.priceType
-
-            const customerPrice = customerPriceType === "org" ? item.data.singlePrice : item.data[customerPriceType]
-            if (customer.connectedDealerId) {
-                const dealer = await ctx.supabase.from("Org").select("*").eq("id", customer.connectedDealerId).single()
-                if (dealer.error) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Failed to get dealer price type",
-                    })
-                }
-                const dealerPriceType = dealer.data.priceType
-                const dealerPrice = dealerPriceType === "org" ? null : item.data[dealerPriceType]
-                const { data: transactionItemDetail, error } = await ctx.supabase.from("TransactionItemDetail").insert({
-                    id: createId(),
-                    customerTransactionId: transaction.data.id,
-                    itemId: i.itemId,
-                    customerPrice: String(customerPrice),
-                    dealerPrice: String(dealerPrice),
-                    quantity: i.totalAdded,
-                    serialNumbers: i.serialNumbers,
-                }).select().single()
-                if (error) {
-                    throw new TRPCError({
-                        code: "BAD_REQUEST",
-                        message: "Failed to create transaction item detail",
-                    })
-                }
-                return transactionItemDetail
-            }
-            const { data: transactionItemDetail, error } = await ctx.supabase.from("TransactionItemDetail").insert({
-                id: createId(),
-                customerTransactionId: transaction.data.id,
-                itemId: i.itemId,
-                customerPrice: String(customerPrice),
-                quantity: i.totalAdded,
-                serialNumbers: i.serialNumbers,
-            })
-            if (error) {
-                throw new TRPCError({
-                    code: "BAD_REQUEST",
-                    message: "Failed to create transaction item detail",
-                })
-            }
-            return transactionItemDetail
-        }))
-
-        return createdTransactionItemDetails
-    }),
     itemCount: protectedProcedure.input(
         z.object({
             storageId: nonEmptyString,
@@ -1246,7 +1013,7 @@ export const itemsRouter = createTRPCRouter({
             }).array().min(1)
         })).mutation(async ({ ctx, input }) => {
 
-            if (!ctx.session.permissions.includes(PERMS.manage_items)) {
+            if (!ctx.session.permissions.includes(PERMS.manage_storage)) {
                 throw new TRPCError({
                     code: "UNAUTHORIZED",
                     message: "You don't have permission to do this!",
